requirements_idx.txt
# Core Dependencies
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.18
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=3.3.0
joblib>=1.2.0

# AI/LLM
google-generativeai>=0.1.0

# Visualization (optional)
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.10.0

# Indonesian Stock Exchange (IDX) Data
idx-scrapper>=0.2.0
# Alternative if the above doesn't work:
# git+https://github.com/antonizer/IDX-Scrapper.git

# Data validation
pydantic>=1.10.0

# HTTP client
aiohttp>=3.8.0

# Progress bars
tqdm>=4.64.0

# Logging and configuration
python-dotenv>=0.19.0